requirements.txt
discord.py
langchain-openai
python-dotenv
openai
langchain
langchain_community
python-telegram-bot